refactor: separate mainSlice
