+++ my-app/src/components/Characters/Characters.tsx
@@ -1,13 +1,8 @@
 import { Mode } from 'helpers/constants/mode';
 import React from 'react';
+import { searchCharactersLoad } from 'redux/asyncThunks';
 import { useAppDispatch, useAppSelector } from 'redux/hooks';
-import {
-  goToFirstPage,
-  goToLastPage,
-  goToNextPage,
-  goToPrevPage,
-  searchCharactersLoad,
-} from 'redux/mainSlice';
+import { goToFirstPage, goToLastPage, goToNextPage, goToPrevPage } from 'redux/mainSlice';
 import Character from './Character';
 import './Characters.css';
