+++ my-app/src/redux/mainSlice.ts
@@ -1,12 +1,15 @@
-import { createAsyncThunk, createSlice, PayloadAction, SerializedError } from '@reduxjs/toolkit';
-import { ICharacter, IName, IState } from 'components/Main/Main';
+import { createSlice, PayloadAction, SerializedError } from '@reduxjs/toolkit';
+import { ICharacter, IState } from 'components/Main/Main';
 import { Mode } from 'helpers/constants/mode';
 import { GenderType, SortingOrder, SortingValues } from 'helpers/constants/sorting';
-import { loadCharacters, searchCharacters } from 'helpers/controllers/getCharacters';
-import { RootState } from './store';
+import {
+  firstCharactersLoad,
+  searchCharactersLoad,
+  searchCharactersThunk,
+  searchNamesThunk,
+} from './asyncThunks';
 
 interface IMainState {
-  names: IName[];
   docs: ICharacter[] | [];
   state: IState;
 }
@@ -14,7 +17,6 @@ interface IErrors {
   errorMessage?: string;
 }
 const initialState: IMainState = {
-  names: [],
   docs: [],
   state: {
     loading: true,
@@ -34,103 +36,6 @@ const initialState: IMainState = {
   },
 };
 
-export const firstCharactersLoad = createAsyncThunk(
-  'main/firstCharactersLoad',
-  async (_, { rejectWithValue, dispatch, getState }) => {
-    const state = getState() as RootState;
-    const { order, sort, gender, limit, page } = state.main.state;
-    const { docs, pages, error, errorMessage, total } = await loadCharacters({
-      page,
-      order,
-      sort,
-      gender,
-      limit,
-      searchValue: '',
-    });
-    if (error) return rejectWithValue({ errorMessage });
-    dispatch(
-      loadCharactersState({
-        ...state.main.state,
-        total,
-        pages: pages || state.main.state.pages,
-        searchValue: localStorage.getItem('searchValue') || '',
-      })
-    );
-    dispatch(addCharacters(docs));
-  }
-);
-export const searchCharactersLoad = createAsyncThunk(
-  'main/searchCharactersLoad',
-  async (_, { rejectWithValue, dispatch, getState }) => {
-    const state = getState() as RootState;
-    const { order, sort, gender, limit, page, searchValue } = state.main.state;
-    const {
-      docs,
-      pages,
-      error,
-      errorMessage,
-      mode,
-      total,
-      page: newPageNumber,
-    } = await loadCharacters({
-      page,
-      order,
-      sort,
-      gender,
-      limit,
-      searchValue,
-    });
-    if (error) return rejectWithValue({ errorMessage });
-    dispatch(
-      loadCharactersState({
-        ...state.main.state,
-        page: newPageNumber,
-        mode,
-        total,
-        pages: pages || state.main.state.pages,
-        searchValue: localStorage.getItem('searchValue') || '',
-      })
-    );
-    dispatch(addCharacters(docs));
-  }
-);
-export const searchCharactersThunk = createAsyncThunk(
-  'main/searchCharactersThunk',
-  async (_, { rejectWithValue, dispatch, getState }) => {
-    const state = getState() as RootState;
-    const { searchValue } = state.main.state;
-    const { docs, error, mode, errorMessage } = await searchCharacters(searchValue);
-    if (error) return rejectWithValue({ errorMessage });
-    dispatch(
-      loadCharactersState({
-        ...state.main.state,
-        mode,
-        searchValue: '',
-      })
-    );
-    dispatch(addCharacters(docs));
-  }
-);
-const timers = {
-  timeout: null as NodeJS.Timeout | null,
-};
-
-export const searchNamesThunk = createAsyncThunk(
-  'main/searchNamesThunk',
-  async (_, { rejectWithValue, dispatch, getState }) => {
-    const state = getState() as RootState;
-    const { searchValue } = state.main.state;
-    if (timers.timeout) clearTimeout(timers.timeout);
-    timers.timeout = setTimeout(async () => {
-      localStorage.setItem('searchValue', searchValue);
-      const data = await searchCharacters(searchValue);
-      const { error, errorMessage } = data;
-      if (error) return rejectWithValue({ errorMessage });
-      const names = data.docs.map(({ name, _id }) => ({ name, id: _id }));
-      dispatch(addNames(names));
-    }, 1000);
-  }
-);
 export type IAction = PayloadAction<
   unknown,
   string,
@@ -166,13 +71,11 @@ export const mainSlice = createSlice({
   name: 'main',
   initialState,
   reducers: {
-    addNames: (state, action: PayloadAction<IName[]>) => {
-      state.names = action.payload;
-    },
     addSearchParams: (state, action) => {
       state.state = { ...state.state, ...action.payload };
     },
     addCharacters: (state, action: PayloadAction<ICharacter[]>) => {
+      state.state.error = false;
       state.docs = action.payload;
     },
     loadCharactersState: (state, action: PayloadAction<IState>) => {
@@ -240,7 +143,6 @@ export const mainSlice = createSlice({
 });
 
 export const {
-  addNames,
   addCharacters,
   loadCharactersState,
   changePage,
