+++ my-app/src/redux/store.ts
@@ -1,11 +1,13 @@
 import { configureStore } from '@reduxjs/toolkit';
 import cardsReducer from './cardsSlice';
 import mainReducer from './mainSlice';
+import namesReducer from './namesSlice';
 
 const store = configureStore({
   reducer: {
     cards: cardsReducer,
     main: mainReducer,
+    names: namesReducer,
   },
 });
 export type RootState = ReturnType<typeof store.getState>;
